backend2/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
openai
dnspython



